Add SEO and LEO optimization to landing page
- Convert app/page.tsx to Server Component with metadata export
- Extract client-side logic to components/landing/LandingPageClient.tsx
- Add generateWebPageSchema() helper for WebPage structured data
- Inject SoftwareApplication and WebPage JSON-LD schemas server-side
- Preserve all existing interactivity and animations in client component
